Moved PicoColorSensor.java into a lib directory
